Pull duplicated code into extra file
