+++ pep2bib.py
@@ -2,62 +2,21 @@
 
 # usage: python3 pep2bib.py .
 
-import datetime
 import glob
 import os
 import re
 import sys
-import time
 from pybtex.database import Entry, BibliographyData
+
+from pep_parsing_helpers import pep_number, pep_creation_dt, first_line_starting_with, parse_authors
 
 BIB_PATH = os.path.join(sys.argv[1], 'peps.bib')
 
 
-def firstline_startingwith(full_path, text):
-    result = None
-    for line in open(full_path, encoding="utf-8"):
-        if result is not None:
-            if not line[0].strip():  # Line begins with whitespace
-                result += line
-            else:
-                return result
-        if line.startswith(text):
-            result = line[len(text):].strip()
-    return None
+name_first_regex = re.compile(r'(.*)<.*>')
+mail_first_regex = re.compile(r'.*\((.*)\)')
+name_only_regex = re.compile(r'(.*)')
 
-
-def pep_creation_dt(full_path):
-    created_str = firstline_startingwith(full_path, 'Created:')
-    # bleh, I was hoping to avoid re but some PEPs editorialize
-    # on the Created line
-    m = re.search(r'''(\d+-\w+-\d{4})''', created_str)
-    if not m:
-        # some older ones have an empty line, that's okay, if it's old
-        # we ipso facto don't care about it.
-        # "return None" would make the most sense but datetime objects
-        # refuse to compare with that. :-|
-        return datetime.datetime(*time.localtime(0)[:6])
-    created_str = m.group(1)
-    try:
-        t = time.strptime(created_str, '%d-%b-%Y')
-    except ValueError:
-        t = time.strptime(created_str, '%d-%B-%Y')
-    return datetime.datetime(*t[:6])
-
-
-def pep_number(full_path):
-    n_str = full_path.split('-')[-1].split('.')[0]
-    try:
-        n = int(n_str)
-    except ValueError:
-        raise Exception("Can't parse pep number %s" % n_str)
-
-    return n
-
-
-name_first_regex = re.compile('(.*)<.*>')
-mail_first_regex = re.compile('.*\((.*)\)')
-name_only_regex = re.compile('(.*)')
 
 months = {
     1: 'jan',
@@ -75,8 +34,7 @@ months = {
 }
 
 
-def clean_authors(authors_str):
-    authors = authors_str.split(',')
+def authors_to_bib(authors):
     cleaned = []
     for author in authors:
         match = name_first_regex.match(author)
@@ -98,12 +56,12 @@ def main():
     # sort peps by date, newest first
     peps_with_dt.sort()
 
-    # generate rss items for 10 most recent peps
     items = {}
     for n, dt, full_path in peps_with_dt:
-        title = firstline_startingwith(full_path, 'Title:')
-        authors = firstline_startingwith(full_path, 'Author:')
-        authors = clean_authors(authors)
+        title = first_line_starting_with(full_path, 'Title:')
+        author_string = first_line_starting_with(full_path, 'Author:')
+        authors = parse_authors(author_string)
+        authors = authors_to_bib(authors)
         url = 'https://www.python.org/dev/peps/pep-%0.4d/' % n
         item = Entry('techreport', [
             ('author', authors),
